Added the solution Container with Max Water
--------------
Brief comment:
--------------

Pick the two bars in such a way so that
offset of the indices * min of the heights
gets maximized i.e.,
Maximize(horizontal_offset * vertical_offset)

Approach:
Pick the 2 bars at the farthest ends.
This guarantees, max horizontal_offset to begin with.
Close from the end (L or R) which has the min height.
This ensures, you're still looking for a bigger height.
Update the max accordingly.